rustc: move the contents of ty::item_path to ty::print.
